# back/docs/swagger.yaml
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  models.ConfiguracioEntrenador:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      duracioSessions:
        type: integer
      entrenadorID:
        type: integer
      id:
        type: integer
      maxAlumnesPerSessio:
        type: integer
      updatedAt:
        type: string
    type: object
  models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  models.HorarisEntrenador:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      desde:
        type: string
      diaSetmana:
        type: integer
      entrenadorID:
        type: integer
      fins:
        type: string
      id:
        type: integer
      updatedAt:
        type: string
    type: object
  models.Reserva:
    properties:
      confirmada:
        type: boolean
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      entrenadorID:
        type: integer
      hora:
        type: string
      id:
        type: integer
      updatedAt:
        type: string
      usuariID:
        type: integer
    type: object
...
  models.Sala:
    properties:
      admin:
        $ref: '#/definitions/models.Usuari'
      adminID:
        type: integer
      codiSala:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      nom:
        type: string
      reserves:
        items:
          $ref: '#/definitions/models.Reserva'
        type: array
      updatedAt:
        type: string
      usuaris:
        items:
          $ref: '#/definitions/models.Usuari'
        type: array
    type: object
  models.SalaInput:
    properties:
      nom:
        type: string
    type: object
  models.SolicitudUnioEntrenador:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      entrenadorID:
        type: integer
      id:
        type: integer
      updatedAt:
        type: string
      usuariID:
        type: integer
    type: object
  models.SolicitudUnioEntrenadorInput:
    properties:
      codiEntrenador:
        type: string
    type: object
  models.SuccessResponse:
    properties:
      data: {}
    type: object
  models.TipusUsuari:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      nom:
        type: string
      updatedAt:
        type: string
    type: object
  models.Usuari:
    properties:
      alumnes:
        items:
          $ref: '#/definitions/models.Usuari'
        type: array
      codiEntrenador:
        type: string
      configuracioEntrenador:
        $ref: '#/definitions/models.ConfiguracioEntrenador'
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      email:
        type: string
      entrenador:
        $ref: '#/definitions/models.Usuari'
      entrenadorID:
        type: integer
      horarisEntrenador:
        items:
          $ref: '#/definitions/models.HorarisEntrenador'
        type: array
      id:
        type: integer
      nom:
        type: string
      password:
        type: string
      reserves:
        items:
          $ref: '#/definitions/models.Reserva'
        type: array
      sales:
        items:
          $ref: '#/definitions/models.Sala'
        type: array
      salesAdministrades:
        items:
          $ref: '#/definitions/models.Sala'
        type: array
      solicitudsUnioEntrenador:
        items:
          $ref: '#/definitions/models.SolicitudUnioEntrenador'
        type: array
      telefon:
        type: string
      tipusUsuari:
        allOf:
        - $ref: '#/definitions/models.TipusUsuari'
        description: 1 = administrador, 2 = basic , 3 = entrenador, 4 = fictici
      tipusUsuariID:
        type: integer
      updatedAt:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/sales:
    get:
      consumes:
      - application/json
      description: Retrieves all the salas from the database.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Sala'
                  type: array
              type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get all salas
      tags:
      - Sales
    post:
      consumes:
      - application/json
      description: Creates a new sala in the database.
      parameters:
      - description: Sala to create
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.SalaInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Sala'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Create a new sala
      tags:
      - Sales
  /api/sales/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a sala from the database.
      parameters:
      - description: ID of the sala to delete
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.SuccessResponse'
            - properties:
                data:
                  type: string
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Delete a sala
      tags:
      - Sales
    get:
      consumes:
      - application/json
      description: Retrieves a sala from the database by its ID.
      parameters:
      - description: ID of the sala to retrieve
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Sala'
              type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get a sala
      tags:
      - Sales
    put:
      consumes:
      - application/json
      description: Updates a sala in the database.
      parameters:
      - description: ID of the sala to update
        in: path
        name: id
        required: true
        type: integer
      - description: Sala to update
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.SalaInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Sala'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Update a sala
      tags:
      - Sales
  /api/sales/salesUsuari:
    get:
      consumes:
      - application/json
      description: Retrieves all the sales of an entrenador from the database.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Sala'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Get all sales of an entrenador
      tags:
      - Sales
  /api/solicitudes/{id}/accept:
    put:
      consumes:
      - application/json
      description: Accepts a pending solicitude in the database.
      parameters:
      - description: ID of the pending solicitude to accept
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.SuccessResponse'
            - properties:
                data:
                  type: string
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Accept a pending solicitude
      tags:
      - Solicitudes
  /api/solicitudes/{id}/decline:
    put:
      consumes:
      - application/json
      description: Declines a pending solicitude in the database.
      parameters:
      - description: ID of the pending solicitude to decline
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.SuccessResponse'
            - properties:
                data:
                  type: string
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Decline a pending solicitude
      tags:
      - Solicitudes
  /api/solicitudes/solicitudsEntrenador:
    get:
      consumes:
      - application/json
      description: Retrieves all the pending solicitudes from the database.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.SolicitudUnioEntrenador'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Get all pending solicitudes
      tags:
      - Solicitudes
  /api/solicituds:
    post:
      consumes:
      - application/json
      description: Creates a new pending solicitude in the database.
      parameters:
      - description: Solicitude to create
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.SolicitudUnioEntrenadorInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.SolicitudUnioEntrenador'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Create a new pending solicitude
      tags:
      - Solicitudes
...
securityDefinitions:
  Bearer:
    description: 'Enter the token with the `Bearer: ` prefix, e.g. "Bearer abcde12345".'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"